[+] Changed Price and salesPrice from Double to BigDecimal

+++ src/main/java/com/bitsmilez/productmicroservice/core/domain/model/Product.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.modelmapper.ModelMapper;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -21,8 +22,8 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private String name;
-    private Double price;
-    private Double salesPrice = null;
+    private BigDecimal price;
+    private BigDecimal salesPrice = null;
     private String img;
     private String description;
     private int stock;
